refactor: move common hook to shared dir
